fix: 시도횟수 검사를 cars 에서 inputview로 이동
