+++ docs/README.md
@@ -50,6 +50,8 @@
 #### 입력
 - [x] 사용자로부터 자동차의 이름들을 입력 받기
 - [x] 사용자로부터 이동 시도 횟수 입력 받기
+  - [x] 유효성 검사
+    - [x] 시도 횟수
 
 #### 출력 
 - [ ] 실행 결과를 출력하기 전 안내 메시지 출력
@@ -69,9 +71,7 @@
   - [x] 입력한 횟수만큼 자동차들에 움직임을 시도하도록 명령
     - [x] 각 자동차들에 움직임을 시도하도록 명령
   - [x] 전체 자동차 이동 현황 반환
-  - [x] 움직임이 끝난 후 우승자 정보 반환
-  - [x] 유효성 검사
-    - [x] 이동 시도 횟수
+  - [x] 움직임이 끝난 후 우승자 정보 반환ㅅ
 
 #### 유틸리티
 - IntGenerator
