+++ src/main/java/racingcar/domain/Cars.java
@@ -7,10 +7,6 @@ import racingcar.util.IntGenerator;
 
 public class Cars {
 
-    private static final int MAXIMUM_TRY_COUNT = 9;
-    private static final String WRONG_TRY_COUNT_ANNOUNCEMENT = "[ERROR] 잘못된 시도 횟수입니다.";
-
-    private static final int MINIMUM_TRY_COUNT = 1;
     List<Car> cars;
 
     public Cars(List<String> carNames) {
@@ -27,18 +23,10 @@ public class Cars {
         this.cars = cars;
     }
 
-    public void repeatRequestMoveBy(int tryCount) {
-        validateTryCount(tryCount);
+    public void requestMoveBy(int tryCount) {
         for (int index = 0 ; index < tryCount ; index++) {
             requestMoveEachCar();
         }
-    }
-
-    private void validateTryCount(int tryCount) {
-        if (MINIMUM_TRY_COUNT <= tryCount && tryCount <= MAXIMUM_TRY_COUNT) {
-            return;
-        }
-        throw new IllegalArgumentException(WRONG_TRY_COUNT_ANNOUNCEMENT);
     }
 
     private void requestMoveEachCar() {
